Fix SPOOKY test error
Jon, in the sprint planning meeting: "I have a feeling something weird
is going to happen with this global css stuff."

Jon, getting unrelated test errors while working on the global css
stuff, ON HALLOWEEN: "That's weird."

--

This is a deceptively small fix for what had turned into a cascade of
errors. First, adding the GlobalCSS component to the MarkOneWrapper
meant that GlobalCSS was being rendered into every test, which is turn
made every test try to load the font packages. This was throwing an
invalid syntax error, because node can't load css. Easy enough, install
`ignore-styles`, add it to the .mocharc.js. Run tests again.

Then I started getting an error about "Cannot read property
'sc-global-3667677342' of undefined". What's that about?  Well, the
GlobalStyleComponent is rendering a `<style>` tag into the
`<head>`, and as a way to ensure that it doesn't add multiple style
tags (especially when using hot module reloading), it keeps a cache
object on the global `window` to track each tag. Apparently, that window
object wasn't being properly initialized in our .mochinit.ts file.

On top of that, `@testing-library/react` has started to automatically
call the cleanup function in an afterEach block when using mocha. Having
the same thing in our .mochainit.ts files was causing cleanup to be
called twice. And somewhere in that first cleanup process it was
removing the cache object from `window`, meaning thatstyled-components
was trying to reference a cache that no long existed.

Cool. So: add jsdom-global/register to the require block of
.mocharc.js, remove the .mochainit file, run tests again. Eveything is
green.

As it happens, I had actually planned to do this in another PR already,
but this forced the issue.