Fix test:watch failures
After my spooky error from Halloween, we found a new bug: Running tests
in watch mode somehow "un-registers" @testing-library's afterEach
cleanup hook on subsequent runs. This is weird and I'll submit a
bugreport to the upstream project, but in the meantime the fix is to
bring back the .mochainit.ts file to add the cleanup hook manually,
while also setting an environment variable to tell testing-library to
not register its own cleanup hook. This should prevent recurrence of
both bugs.