Create overview, guidelines, and sections
To integrate Vittorio's work on the interaction guidelines, I've added
some logic to parse through a `guidelines` directory and add any
markdown files in there to the styleguide output, under a "Guidelines"
heading.

Expanding on that, I configured it to generate separate pages for each
sub-section of components, and to add our README and CONTRIBUTING
documents as the first entries in the library. I think this should make
the whole document a little more user friendly.