Update color specimens
Because some of the "medium" values are brighter than the others, the
previous logic of "use light text if the background contains dark"
doesn't hold. Instead I added a getFontColor function that determines
what color the text should be.

In practice this would probably be set manually for any given component.
It's only because I wanted to generate these samples programmatically
that this is an issue here.