Replace relative module paths
When compiling for consumption in other projects, typescript libraries are typically transpiled (in our case using `tsc`) but not actually bundled using webpack or some other module bundler. In turn this means that relative paths like `mark-one` persist in the compiled output.

The downstream consuming application then attempts to resolve this path itself which results in the bundler(webpack) attempting to import uncompiled TSX typescript assets, instead of their compiled JavaScript counterparts which results in webpack errors/warnings about not compiling dependencies with your app. 

In general, the best solution is to use relative paths so that the bundler can resolve the dependencies of each file in the correct context.