Postinstall script to transpile components
A pre-publish hook currently exists to transpile the typescript into javascript before an npm publish, however this isn't useful to us right now as the app is not currently published on npm, but is just published on GitHub. 

Therefore, a postinstall script is also needed to transpile the TSX when npm installs `mark-one` form GitHub. 

Eventually, when the library is released and published on NPM this should probably be removed/reverted.