More explicit tag spec
I misunderstood the conditionals section of the travis docs, so the
previous config was not actually building the npm package. This update
should be a little more specific, and should only build the package when
there's a semver tag (e.g. "v0.0.1"). We'll need to adjust the regex if
we want to publish pre-release packages in the future.

https://docs.travis-ci.com/user/deployment/npm/
https://docs.travis-ci.com/user/conditions-v1