Create Modal Header
Adds a Header bar to the modal, which renders any `children` text inside
a slightly modified `<SectionHeading>` component (removes the bottom
border). Also accepts an optional `closeButtonHandler` prop: if present,
it renders a close button in the top corner; if not, it doesn't render
the button and closing functionality might be handled in another way.