Refactor Table* components
Previously, the Table* components were creating functional components
that were just getting the theme, passing props to the styled
component, and passing through children. This works, but does make it a
bit harder to add props to the components, since each prop has to be
individually passed through to the underlying styled component.

Now we're exporting the components wrapped in the `withTheme` Higher
Order Component. This is achieving the same final result as before, but
no longer requires us to pass props though. In my opinion, this is how
we should be handling any component that is just passing props/children
through to an underlying styled-component.