Test ternary in :hover on TableRow
I thought this was going to be as easy as using `fireEvent` to send a
mouseOver event, then looking at the computedStyles the way we have with
other elements. Of course it was more difficult that that.

Short answer: There's no way to view/manipulate pseudo-classes with
Javascript, in the sense of activating the :hover attribute of an
element. What we have to do instead is analyze the stylesheets in the
document and look at for a :hover variant on the styled-components
className and see what the rules are in there. This is a bit stupid,
and comes close to being a case of testing styled-components, but at
least it gets that last bit of code coverage squared away.