Continued adventures in Typescript
Sigh. It turns out that just declaring the Type isn't quite enough for
styleguidist, so I had to also add a _different_ annotation to the const
for the component, which makes the props & methods show up correctly in
the styleguide.

Along the way I also fixed some of the prop types to correctly handle
cases where a component could have one or more of multiple components as
children.