Use string for {row,col}Span
While the value passed in is almost always going to be a number, using
the `number` type breaks the expected behavior for primitive props. In
short, with `number` the usage would have to be:

```tsx
<TableHeadingCell colSpan={2}>Heading</TableHeadingCell>
```

While with `string` the usage would be:

```tsx
<TableHeadingCell colSpan="2">Heading</TableHeadingCell>
```

A very minor difference, but I feel like it makes for a better
development experience when our react code reads more like html.