Replace required property with isRequired and add aria-required property
aria-required property is needed for accessibility purposes to indicate whether a TextInput field is required. We are using an isRequired property instead of the built in required property so that we can do custom error handling ourselves. When the isRequired property is true, a red asterisk will appear next to the label of the input field as a signal to the user that that field is required in order to submit the form.