Update Modal and ModalHeader components to include a forwardRef property that is used to trap the focus in the modal
Add optional forwardRef properties to the Modal and ModalHeader. These props specify the refs of the modal, which are used to direct the focus. When the user clicks on the button to open the modal, the focus should be directed to the Modal Header. Add logic to redirect the focus to the first focusable element in the modal when the modal is visible too so that the focus cannot be tabbed to the background elements.