Fix "unsafe any" errors in Buttons
Accessing object values with variable names results in an eslint error
because it's not obvious what the type of the value would be. We can
sidestep this issue by using our fromTheme function, since we can
provide a concrete return type. The syntax is a bit ugly because
fromTheme returns a function that needs to be called with a props
object.