Remove onOpen and onClose props from Modal
Originally we were using the useEffect hook to call the onOpen and
onClose handlers without watching those values for changes. This is
incorrect, but watching those values for changes could result in an
infinite loop situation if the handlers were relying on hooks in the
parent context, since those underlying values would not be exposed
within the useEffect hook itself.

The simplest mitigation is to simply remove the onClose and onOpen
props, and let any additional effects be handled by the parent component
when the `isVisible` value is changed. I've updated the example file to
demonstrate.

Reference: https://reactjs.org/docs/hooks-faq.html#is-it-safe-to-omit-functions-from-the-list-of-dependencies