Handle aria-errormessage and aria-invalid
When there is an errorMessage prop passed to the component, the
aria-invalid attribute should exist on the Input/Dropdown components,
and when there is not one the prop should not exist at all.
Additionally, we need to set the aria-errormessage value for the input
to the id of the error message, which means we need to create an id for
the errorMessage by appending `-error`.

Additionally, we should make the `id` prop required, as it's used to
associate the label, input, and error.