Rewrite fromTheme as recursive
This started as an effort to rewrite the fromTheme function by calling
`Array.reduce()` on the list of property names, but I was having trouble
safely typing the values. This revised form makes use of a recursive
function to descend through the Theme tree, using Typescript's index
types. Because we want to be able to handle arbitrary depths of
recursion, we can only safely type the argument to `fromTheme`. This
means we still have to give up some of the benefit of using typescript
(i.e. compiler errors when accessing an invalid value). It may be
possible to create more specific types using conditional types, but I
think we can save that for a future enhancement.