Overhaul typing of theme
Previously, we were relying on literal types for most of the objects in
our theme interface. This made it much harder to properly typecheck,
especially in the fromTheme function. To improve the type safety, we're
not declaring a union type for each level of the theme, and asserting
that each key exists in that union type.

The color.background section (L67-L72) proved to be the most difficult,
as the values there were either strings or ColorRange. A simply union
type (ie. `[key: string]: string | ColorRange`)wasn't sufficient, since
that required every value to be assignable to both types. Instead, we
can use a key index assertion that every key is either a VARIANT or
a key in the ColorRange, then set the value to conditionally be a
ColorRange or a string depending on what the key is.

We're also exposing a ValidThemeValues type, which is just a big union
of all the possible theme keys.