Enable proper typings in fromTheme
Now that we have a strongly typed theme, we can enhance our existing
fromTheme function to check the values for type safety. This did require
a concession that the function can now only go four levels deep, instead
of infinitely deep, but this should not be an impediment in most cases,
since the normal functional style is still a valid option.

Because this function is now type-safe, we no longer need to test that
it correctly handles passing in wrong values.