Split sublevel theme types into new file
The existing method of declaring all the theme types as part of the
styled-components namespace was breaking when the library was published,
since none of those themes were being exported and thus were missing
when Theme/utils.ts tried to import them. This moves those sub-levels
into the ThemeTypes.ts file (which will be compiled/exported), then
redclares the ValidThemeValues as part of the utils module itself.