Update the formatting of input elements so that checkbox and radio buttons do not appear in the center when labelPosition is set to POSITION.TOP
Originally, the checkbox and radio buttons appeared close to the center of its parent due to the way we were styling with CSS grid. This presents issues when the start of the elements are not aligned in a form. As a fix, the grid-template-columns was set so that columns will shrink if there's no content in it.