Create ExternalLink component
Our existing Link component is a wrapper around ReactRouter's Link, so
we can only use it for navigation within the app. This ExternalLink is
just a `styled.a` with CSS to match the Link, so that we can create
links to external pages that look similar.

The specific use case for this is the "log out" link from #246