Add a generic type for the Mark One Props interface
I got a type error in the build - ' Type 'RefObject<HTMLElement>' is not assignable to type 'RefObject<HTMLButtonElement>'' - To fix this, I updated the interface to accept a type that extends the HTMLElement. Depending on what the component is, it will extend the indicated type. T is the underlying element you're going to reference