Expose disabled and onClick properties for Dropdown component
For the Course Admin modal, we have an area section where the user either creates a new area or selects an existing one. When the user is creating a new area, we would like the dropdown where they would have been selecting an existing area to be disabled. I thought it would also be useful to expose the onClick property for the case where you have a group of radio buttons and want to select a radio button when the dropdown for a particular radio button selection is clicked.