Allow deployment on all branches
By default, Travis will only run a deploy command on `master`, and we
want this to run on `default` and `main`. Since we already have a
conditional check in the 'stages' section of the yaml, as well as a check
inside the deploy script itself, we should be safe to allow this
deploy stage to run on any branch.