Add a data test id to the form for testing purposes
To test that the form component is truly being rendered, I added an optional data-testid to the form so that we can check for this in the test. Originally, I was checking to see if the error message within the text input that was in the form was being rendered, but this is not truly testing if the form itself is being rendered.