Fix bugs in Table Cell tests
The cell containing the text content of "2" was updated to have a variant of VARIANT.MEDIUM, so the styles of that cell wer changed - the tests checking for the case when the variant is not specified were updated to get the style of a different cell instead. The font weight of the VARIANT.MEDIUM and VARIANT.NEGATIVE cells is the bold font weight, not the medium font weight.