Add ButtonDropdown example
This example demonstrates a ButtonDropdown with differing values and labels. When the user clicks the button dropdown, it reveals the div that outlines the options. If the user initially clicks the button dropdown and clicks it a second time, it will hide the menu. When the user chooses a selection, the value of the option will be alerted.