Separate ButtonDropdownMenuItem from ButtonDropdownMenu
The purpose of this change is to allow users to be able to customize the options that they use as demonstrated in the markdown file instead of being prescriptive of the options always just containing a value and label. This allows the user to be more flexible with the onClick handler as well.