Create generic list components
These components are variations of the TableCellList and
TableCellListItem, with an "ordered" prop added to the List component to
switch between a ul and ol element. This doesn't affect the appearance
of the list, but it still important semantically.

Additionally, the borders between the list children have been moved to
the parent List element, which provides just a bit more flexibility if
we ever want to create multiple list components.