Add role and aria-live properties to ModalMessage
In most cases, I would anticipate the role and aria-live values would be 'alert' and 'assertive, respectively. Since this may not always be the case, I added the props role and ariaLive so that the user can change the values but also added default values for both of these properties.