Create the Combobox
This is sort of a variation on our Dropdown component, which uses a
typeahead/autocomplete interface to choose from among a list of options.
I've tried to mirror the props interface of the TextInput as much as
possible and to reuse the existing lower level components. This did
require re-wrapping a few of them to apply slightly different styling
(mainly in positioning/spacing), but the overall look and feel should be
consistent.