Allow custom filtering
The initial implementation only allowed for a case-insensitive search of
the user input against the option label. This will probably cover most
circumstances, but when we implement this Combobox to add instructors
we will likely want to allow for a broader search, e.g. to match
against the first name or last name, or match substrings.

That initial implementation is now the default search for those cases
where it's all we need.