Create a GlobalMessage component
The GlobalMessage component is a styled div component that contains a button that will handle closing the message along with the message contents (via the children prop). I used CSS grid to handle the formatting of the text and button so that the text is centered and the button is right aligned.