Remove isLabelVisible prop for required example
It makes sense to show the label for a required input field so that the
required asterisk indicating the field's required status is visible. Otherwise
it just looks like any other input field in the examples.

Since `isLabelVisible` defaults to `true` we can just remove this prop entirely
from the example