Create Popover component
This renders a small informational pop-up box whose visibility can be
controlled via an isVisible prop. When set to visible, it will appear
directly above and to the right of its location in the DOM. Optional
xOffset and yOffset parameters can be used with positive and negative
values to move the popup into different positions.

I did explore adding logic to prevent the popup from appearing outside
the visible portion of the page, but ulitmately decided that was adding
unnecessary complexity. That feature could be enabled at implementation
time using a callback Ref, as in the React Docs example for measuing a
DOM node[1]

[1]: https://reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node